+++ docs/source/changelog.md
@@ -6,7 +6,8 @@
   * Rename `TableReference` to `ExternalTableReference`
   * Add support for `ExternalTableReference` to point to tables in external (i.e. not managed by `pipedag`) schemas. 
   * Remove support for `ExternalTableReference` that points to table in schema of current stage. I.e. `ExternalTableReference` can only point to tables in external schemas.
-- Support code based configuration (see simplified README.md example without config file and without docker-compose)
+- Support code based configuration (see get_basic_pipedag_config() in README.md example without config file and without docker-compose)
+- Added NoBlobStore in case you don't want to provide a directory that is created or needs to exist
 - Fix polars import in `pyproject.toml` when using OS X with rosetta2
 
 ## 0.6.10 (2024-02-29)